Tidy terminal runtime live message handling
